Move the remainder of the spa-specific stuff
.. update docs, tests
