+++ src/argus/auth/views.py
@@ -1,5 +1,3 @@
-from django.contrib.auth import logout
-from django.conf import settings
 from django.db import transaction
 
 from drf_spectacular.utils import extend_schema, extend_schema_view
@@ -7,7 +5,7 @@ from rest_framework import generics
 from rest_framework.authtoken.models import Token
 from rest_framework.authtoken.views import ObtainAuthToken
 from rest_framework.exceptions import ValidationError
-from rest_framework.permissions import IsAuthenticated, AllowAny
+from rest_framework.permissions import IsAuthenticated
 from rest_framework.response import Response
 from rest_framework.views import APIView
 
@@ -31,24 +29,6 @@ class ObtainNewAuthToken(ObtainAuthToken):
             pass
         token = Token.objects.create(user=user)
         return Response({"token": token.key})
-
-
-class LogoutView(APIView):
-    permission_classes = []
-
-    @extend_schema(request=None, responses={"200": None})
-    def post(self, request, *args, **kwargs):
-        "Log out the logged in user"
-        user = request.user
-        if hasattr(user, "auth_token"):
-            user_token = request.user.auth_token
-            user_token.delete()
-        # Log out from session
-        logout(request)
-
-        response = Response()
-        response.delete_cookie(settings.ARGUS_TOKEN_COOKIE_NAME)
-        return response
 
 
 class CurrentUserView(APIView):
