+++ src/argus/site/settings/base.py
@@ -270,10 +270,6 @@ NOTIFICATION_SUBJECT_PREFIX = "[Argus] "
 # Don't spam by accident
 SEND_NOTIFICATIONS = get_bool_env("ARGUS_SEND_NOTIFICATIONS", default=False)
 
-COOKIE_DOMAIN = get_str_env("ARGUS_COOKIE_DOMAIN", None)
-
-ARGUS_TOKEN_COOKIE_NAME = "token"
-
 # 3rd party settings
 
 # Python social auth
