+++ src/argus/site/settings/dev.py
@@ -1,4 +1,5 @@
 import logging.config
+import os
 
 from dotenv import load_dotenv
 
@@ -6,12 +7,12 @@ load_dotenv()
 
 from . import get_bool_env, get_str_env  # noqa: E402
 from .base import *  # noqa: E402, F403
-from argus.spa.settings import *  # noqa: E402, F403
 
+os.environ.setdefault("ARGUS_SPA_COOKIE_DOMAIN", "localhost")
+from argus.spa.spa_settings import *  # noqa: E402, F403
 
 DEBUG = get_bool_env("DEBUG", True)
 TEMPLATES[0]["OPTIONS"]["debug"] = get_bool_env("TEMPLATE_DEBUG", True)  # noqa: F405
-
 
 # SECURITY WARNING: keep the secret key used in production secret!
 SECRET_KEY = get_str_env("SECRET_KEY", "secret-secret!")
