+++ src/argus/spa/__init__.py
@@ -1,0 +1,7 @@
+"""
+This module and its submodules are necessary for the set up and use of the
+legacy single page application (SPA) REACT frontend which can be found at
+https://github.com/uninett/Argus-frontend
+
+See docs/reference/react-frontend.rst
+"""
