+++ src/argus/spa/spa_settings.py
@@ -1,0 +1,31 @@
+from argus.site.settings import get_str_env
+
+
+FRONTEND = "spa"
+
+LOGIN_URL = "/login/"
+LOGOUT_URL = "/logout/"
+LOGIN_REDIRECT_URL = "/"
+LOGOUT_REDIRECT_URL = "/"
+
+ARGUS_SPA_TOKEN_COOKIE_NAME = "token"
+SPA_COOKIE_DOMAIN = get_str_env("ARGUS_SPA_COOKIE_DOMAIN", required=True)
+
+# PSA for login
+
+_SOCIAL_AUTH_DATAPORTEN_KEY = get_str_env("ARGUS_DATAPORTEN_KEY")
+_SOCIAL_AUTH_DATAPORTEN_SECRET = get_str_env("ARGUS_DATAPORTEN_SECRET")
+
+SOCIAL_AUTH_DATAPORTEN_EMAIL_KEY = _SOCIAL_AUTH_DATAPORTEN_KEY
+SOCIAL_AUTH_DATAPORTEN_EMAIL_SECRET = _SOCIAL_AUTH_DATAPORTEN_SECRET
+
+SOCIAL_AUTH_DATAPORTEN_FEIDE_KEY = _SOCIAL_AUTH_DATAPORTEN_KEY
+SOCIAL_AUTH_DATAPORTEN_FEIDE_SECRET = _SOCIAL_AUTH_DATAPORTEN_SECRET
+
+AUTHENTICATION_BACKENDS = [
+    "argus.spa.dataporten.social.DataportenFeideOAuth2",
+    "django.contrib.auth.backends.RemoteUserBackend",
+    "django.contrib.auth.backends.ModelBackend",
+]
+
+ROOT_URLCONF = "argus.spa.urls"
