+++ tests/auth/test_auth.py
@@ -1,6 +1,3 @@
-from datetime import timedelta
-
-from django.conf import settings
 from django.contrib.auth import get_user_model
 from django.urls import reverse
 from rest_framework import status
@@ -8,6 +5,8 @@ from rest_framework.authtoken.models import Token
 from rest_framework.test import APIClient, APITestCase
 
 from argus.auth.factories import AdminUserFactory, PersonUserFactory
+
+from . import assemble_token_auth_kwarg, expire_token
 
 
 User = get_user_model()
@@ -23,36 +22,11 @@ class APITests(APITestCase):
 
         self.superuser1_client = APIClient()
         self.superuser1_token = Token.objects.create(user=self.superuser1)
-        self.superuser1_client.credentials(**self.assemble_token_auth_kwarg(self.superuser1_token.key))
+        self.superuser1_client.credentials(**assemble_token_auth_kwarg(self.superuser1_token.key))
 
         self.normal_user1_client = APIClient()
         self.normal_user1_token = Token.objects.create(user=self.normal_user1)
-        self.normal_user1_client.credentials(**self.assemble_token_auth_kwarg(self.normal_user1_token.key))
-
-    @staticmethod
-    def assemble_token_auth_kwarg(token_key: str):
-        return {"HTTP_AUTHORIZATION": f"Token {token_key}"}
-
-    @staticmethod
-    def expire_token(token: Token):
-        # Subtract an extra second, just to be sure
-        token.created -= timedelta(days=settings.AUTH_TOKEN_EXPIRES_AFTER_DAYS, seconds=1)
-        token.save()
-
-    def test_logout_deletes_token(self):
-        logout_path = reverse("v1:auth:logout")
-
-        def assert_token_is_deleted(token: Token, user: User, client: APIClient):
-            self.assertTrue(hasattr(user, "auth_token"))
-            self.assertEqual(client.post(logout_path).status_code, status.HTTP_200_OK)
-
-            user.refresh_from_db()
-            self.assertFalse(hasattr(user, "auth_token"))
-            with self.assertRaises(Token.DoesNotExist):
-                token.refresh_from_db()
-
-        assert_token_is_deleted(self.normal_user1_token, self.normal_user1, self.normal_user1_client)
-        assert_token_is_deleted(self.superuser1_token, self.superuser1, self.superuser1_client)
+        self.normal_user1_client.credentials(**assemble_token_auth_kwarg(self.normal_user1_token.key))
 
     def _successfully_get_auth_token(self, user: User, user_password: str, client: APIClient):
         auth_token_path = reverse("v1:api-token-auth")
@@ -86,7 +60,7 @@ class APITests(APITestCase):
         def assert_token_expires_and_is_deleted(user: User, token: Token, client: APIClient):
             self.assertEqual(client.get(some_auth_required_path).status_code, status.HTTP_200_OK)
 
-            self.expire_token(token)
+            expire_token(token)
 
             self.assertEqual(client.get(some_auth_required_path).status_code, status.HTTP_401_UNAUTHORIZED)
             user.refresh_from_db()
@@ -96,33 +70,6 @@ class APITests(APITestCase):
 
         assert_token_expires_and_is_deleted(self.normal_user1, self.normal_user1_token, self.normal_user1_client)
         assert_token_expires_and_is_deleted(self.superuser1, self.superuser1_token, self.superuser1_client)
-
-    def test_can_get_auth_token_after_deletion_or_expiration(self):
-        logout_path = reverse("v1:auth:logout")
-        some_auth_required_path = reverse("v1:auth:current-user")
-
-        def assert_unauthorized_until_getting_auth_token(user: User, user_password: str, client: APIClient):
-            self.assertEqual(client.get(some_auth_required_path).status_code, status.HTTP_401_UNAUTHORIZED)
-
-            client.credentials()  # clears credentials
-            response = self._successfully_get_auth_token(user, user_password, client)
-            client.credentials(**self.assemble_token_auth_kwarg(response.data["token"]))
-
-            self.assertEqual(client.get(some_auth_required_path).status_code, status.HTTP_200_OK)
-
-        def assert_can_get_auth_token_after_deletion_and_expiration(user: User, user_password: str, client: APIClient):
-            client.post(logout_path)
-            assert_unauthorized_until_getting_auth_token(user, user_password, client)
-
-            self.expire_token(Token.objects.get(user=user))
-            assert_unauthorized_until_getting_auth_token(user, user_password, client)
-
-        assert_can_get_auth_token_after_deletion_and_expiration(
-            self.normal_user1, self.normal_user1_password, self.normal_user1_client
-        )
-        assert_can_get_auth_token_after_deletion_and_expiration(
-            self.superuser1, self.superuser1_password, self.superuser1_client
-        )
 
     def test_get_current_user_returns_correct_user(self):
         current_user_path = reverse("v1:auth:current-user")
